+++ TrainersSchoolingSystem/Models/DTOs/PaidFeeViewModel.cs
@@ -10,8 +10,8 @@ namespace TrainersSchoolingSystem.Models.DTOs
         public int PaidFeeId { get; set; }
         public Nullable<int> StudentId { get; set; }
         public string Description { get; set; }
-        public Nullable<int> CalculatedAmount { get; set; }
-        public Nullable<int> ReceivedAmount { get; set; }
+        public Nullable<decimal> CalculatedAmount { get; set; }
+        public Nullable<decimal> ReceivedAmount { get; set; }
         public Nullable<System.DateTime> PaymentDate { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
         public Nullable<int> CreatedBy { get; set; }
